Fix color problem and fix time step to display in BBox information

+++ lidarFrame.cxx
@@ -1,6 +1,6 @@
 #include "lidarFrame.h"
 
-LidarFrame::LidarFrame(int id, vtkSmartPointer<vtkPolyData> points,
+LidarFrame::LidarFrame(size_t id, vtkSmartPointer<vtkPolyData> points,
                        const Eigen::Isometry3d &pose, double time)
   : m_id(id), m_points(vtkSmartPointer<vtkPolyData>::New()), m_pose(pose), m_timestamp(time)
 {
